Check if we have a user in is_elevated
On start-up when we define the properties is_elevated gets called even
if we don't have a current user. This violates the assert in GetUserId()
and errors out in builds that are compiled with assertions enabled. The
logic still works with invalid OIDs but it's better to handle that case
explicitly.